{"schema":"decodex/commit/1","summary":"Extract scroll trace schema model owner","authority":"manual"}
